requirements.txt
click==6.7
cycler==0.10.0
enum-compat==0.0.2
eventlet==0.20.1
Flask==0.12
greenlet==0.4.11
h5py==2.6.0
itsdangerous==0.24
Jinja2==2.9.3
Keras==1.2.0
MarkupSafe==0.23
matplotlib==1.5.3
numpy==1.11.3
olefile==0.44
pandas==0.19.2
Pillow==4.0.0
protobuf==3.1.0.post1
pygame==1.9.2
pyparsing==2.1.10
python-dateutil==2.6.0
python-engineio==1.1.0
python-socketio==1.6.2
pytz==2016.10
PyYAML==3.12
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
sklearn==0.0
tensorflow-gpu==0.12.1
Theano==0.8.2
Werkzeug==0.11.15
